pyproject.toml
[tool.poetry]
name = "bordereau-generator"
version = "0.1.0"
description = ""
authors = ["skyfion75 <152012451+skyfion75@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.28.2"
python-docx = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
